scripts/requirements.txt
pynostr>=0.6.0
websocket-client>=1.6.0
